+++ DX11Starter/DX11Starter/DX11Starter.vcxproj
@@ -159,13 +159,11 @@
   <ItemGroup>
     <ClCompile Include="FilePathHelper.cpp" />
     <ClCompile Include="Game.cpp" />
-    <ClCompile Include="GameUI.cpp" />
     <ClCompile Include="Main.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="FilePathHelper.h" />
     <ClInclude Include="Game.h" />
-    <ClInclude Include="GameUI.h" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
